+++ src/main/java/ru/yandex/money/gradle/plugins/gradleproject/GradleProjectPlugin.java
@@ -4,11 +4,10 @@ import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.gradle.plugin.devel.plugins.JavaGradlePluginPlugin;
 import org.gradle.util.VersionNumber;
-import ru.yandex.money.gradle.plugin.architecturetest.ArchitectureTestPlugin;
-import ru.yandex.money.gradle.plugins.backend.build.JavaModulePlugin;
 import ru.yandex.money.gradle.plugins.javapublishing.JavaArtifactPublishPlugin;
 import ru.yandex.money.gradle.plugins.library.git.expired.branch.GitExpiredBranchPlugin;
 import ru.yandex.money.gradle.plugins.task.monitoring.BuildMonitoringPlugin;
+import ru.yoomoney.gradle.plugins.moduleproject.ModuleProjectPlugin;
 import ru.yoomoney.gradle.plugins.release.ReleasePlugin;
 
 import java.util.Arrays;
@@ -23,18 +22,17 @@ import java.util.Collection;
 public class GradleProjectPlugin implements Plugin<Project> {
 
     private static final Collection<Class<?>> PLUGINS_TO_APPLY = Arrays.asList(
-            JavaModulePlugin.class,
+            ModuleProjectPlugin.class,
             JavaArtifactPublishPlugin.class,
             ReleasePlugin.class,
             GitExpiredBranchPlugin.class,
-            BuildMonitoringPlugin.class,
-            ArchitectureTestPlugin.class
+            BuildMonitoringPlugin.class
     );
 
     @Override
     public void apply(Project project) {
-        if (VersionNumber.parse(project.getGradle().getGradleVersion()).compareTo(VersionNumber.parse("4.10.2'")) < 0) {
-            throw new IllegalStateException("Gradle >= 4.10.2 is required");
+        if (VersionNumber.parse(project.getGradle().getGradleVersion()).compareTo(VersionNumber.parse("6.4.1'")) < 0) {
+            throw new IllegalStateException("Gradle >= 6.4.1 is required");
         }
 
         project.getPluginManager().apply(JavaGradlePluginPlugin.class);